db-migrate: Removing database name override
To test out db-migrate I was setting a lambda environment variable
to use a different database. It is time to apply the changes to the
default database. On a related note, in Secrets Manager I have changed
the default database from "config" to "main", so the database that will
be updated by db-migrate is "main".