postgres: Fixed id generation concurrency issue
To improve efficiency, I was selecting/updating the object sequences
record in a single statement when generating an id for a record. However
I saw that this could lead to some issues with concurrencey (it presented
as duplicate primary keys) so I now select the record for update, pull the
values I need, then update it in a second statement. I tested this by
executing a number of parallel inserts both before and after the change
and the fix appears to address the problem.