json-schema: JSON representation of an organizations db schema
Adding a new module that will help make changes to an organization's
database schema. So far the module includes the following functionality:
1. Validate whether a JSON object is a valid schema representation
2. Create a JSON schema object from the Postgres information_schema
   tables
3. Create a JSON schema object from our own object/object_field tables

The next steps will be to be able to compare JSON schema objects to
see how they differ, and then to generate changes that will change the
db schema from one representation to another.