tests: moving data-abstration and json-schema tests to new projects
Since json-schema depends on data-abstraction (for DbObject and
DbObjectField classes), but for tests data-abstraction also
depended back on json-schema (for DataObjectGenerator class). This
circular dependency was creating issues so I moved the tests out
into their own projects, both of which require both packages.

The install-local module does not support peer dependencies (which
would be great to help ensure each module is always installed alongside
the other), but since these are private modules I will always be in
control of it should be manageable.