database-test: Adding swappable database-test module
I am adding a new swappable database module for testing. This
module has a new interface defined in database-interface, and is
meant to serve as a database mock while running unit tests. The
interface for now is super simple. In addition to the standard
query() method, it also offers setResponse(obj) to set whatever
response is desired from the db and getLastQuery() so the exact
format of the previously run query can be viewed.