data-abstraction: constructor properties can prepopulate DbObjects
Added an optional properties parameter to the DbObject constructor to allow those properties to be set during instantiation. The properties just get passed through to the "set" method so they are all checked at runtime to verify the properties are valid. This will be very useful for unit testing in particular.

I also update data_object_generator to support this as well, and created a test to verity it is working.