json-schema: Adding configurable test_scope to generated class names
In general, each test file generates its own schema based on its own needs. However, DbObject indexes the dbFields based on class name, so we need some way to differentiate the different instances that get created. The optional "test_scope" property on the table gets appended to the table name to deal with this. It is expected that it will be set uniquely based on the name of the test file.