data-abstraction: improvements to Id class
I've made a couple changes to the Id class:
1) It no longer extends the String class. I have seen recommendations against subclassing this way, and it's not actually useful (I was hoping it would help convert to strings when appropriate, but it doesn't)
2) I implemented [Symbol.toPrimitive](). This helps (in some cases) convert the Id to a string in the proper context. When testing this, I actually found different behavior (in terms of calls to toString and toPrimitive) between templated strings (`blah ${id}`) and string concatenation ("blah " + id).

I am still concerned that keeping this class around (or at least allowing instances of it to be instantiated) will continue to lead to a number of gotchas down the road. In the future I may just use static members on the class for validation, etc but all Id's will in fact be primitive strings.