tooling: Adding --build-in-source flag to samconfig.toml file
In order to get install-local working inside of my sam projects, I need to build in the actual folder instead of copying the files out to a temp folder (in which case the the local paths are no longer correct). Sam supports this by supplying --build-in-source to the sam build tool (i.e. "sam build --build-in-source MyProject"). But sam also allows for build parameters to be specified in the automatically loaded configuration file samconfig.toml. In order to do this, the leading dashes must be removed and the internal dashes must be converted to underscores. Hence we have "build_in_source = true".