database-direct: Removed try/catch around query execution
Catching the exception was causing unexpected behavior elsewhere. I now let the error propagate up. This could have also been done by throwing the error again or throwing a new one. This may be a better long-term option but for now I am just letting the original error travel up the stack so that it's clear the query did not succeed.