db-migrate: Changing uid length from 7 to 8 characters
When creating ids I did a little thinking and decided to change the uid portion of the id to be 8 characters instead of just 7. With 7, there were ~78B uids possible, but by going up to 8 that increases to 2.8T. While not ridiculously large, after a few calculations it is hard to imagine ever exhausting this number on a single table.