db-migrate: Adding sequence_number column to object_sequences table
Adding a new sequence_number column that will keep track of how far through the address space a given sequence is. Without it the sequence generation was functional but it would be very difficult to see how many ids had been generated (or more importantly how many future ids can yet be generated).

The column starts at zero and increments by 1 every time the sequence is used. There is a CHECK constraint to ensure it remains below the value stored in the "max" column.

There were also a couple changes such that when updating the object_sequences table the row is locked before it is updated. This had been done for uids previously but has been added to org key generation as well as setting object prefixes.