db-migrate: Adding organization_key column to objects table
I have decided to add organization_key to the objects table to guarantee referential integrity. It is redundant (which also means it can be auto populated), but I decided the tradeoffs were worth it. The reason behind it is that I want the object_sequences table to use the org key instead of the org id. The sequences table is there to automatically generate id's, and the org key is used for this. Yes, we could do a lookup back to the organization table if we had the id stored in the sequences table, but that seems less elegant to me than duplicating the org key in the objects table. We were simulating referential integrity before by having foreign keys to both the object and organizations table, but there was no way to close the loop and ensure that the object record pointed to the same organization record referenced in the sequences table.