db-migrate: Adding sql_type field to object_fields table
It will be helpful down the road to know the intention of a field (e.g. "id") vs. the true representation ("text") in sql. I would like these to be picklists since they have a limited set of fields, but I'm not exactly sure of the implementation yet. Better yet "sql_type" would be a dependent picklist, tied to the values in the "type" field.