db-migrate: Adding default_type column to object_fields table
I was trying to create a new object by inserting records into objects and object_fields, but realized I couldn't have a created_at field that is automatically set (with the now() function) because there was no distinction between a text value and a formula. I have now added a "default_type" column to the object_fields table that
1) must be set if default_value is set
2) must be null if default_value is set
3) can be either the same as the "sql_type" value, or the word "formula"

When automatically generating the column by inserting an object_field with a default value, the default value is only quoted if the default_type is equal to "text".